refactor(examples): splitting src of PieChart example; split
